+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Raise error from yamlet renderer when an included file does not exist
+- Experimental inventory feature
 
 ## [1.10.0] - 2022-09-03